Add functionality
We had the problem, that the DateTimeField provided by Flask wasn't working properly and the DateTimeLocalField wasn't accepted by the database. Therefore, we had to change the way the datetime was inserted: now it gets pushed to the server as a time object and a date object, which can be combined via datetime.datetime.combine(element1, element2).

The server understands the combination and even can fill the correct part into the fields when editing the orders(The GET request works properly therefore)

Next step is to add the functionality to change the "start_at" value, because you are currently only able to set it once

-new styling of datetime in representation with strftime:D